Split listOps into more sensible code units. Walk back 'reducer' idea, for now.
