+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "@auth0/auth0-react": "^1.12.1",
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
     "@iconify/react": "^3.2.1",
     "@material-ui/core": "^4.10.0",
     "@material-ui/icons": "^4.9.1",
